refactor(#2, #23, #25, #30) : QueryService, CommandService 분리 + 에러코드 도메인별로 정리
