+++ src/main/java/com/example/picket/domain/ticket/service/TicketQueryService.java
@@ -1,0 +1,41 @@
+package com.example.picket.domain.ticket.service;
+
+import com.example.picket.common.exception.CustomException;
+import com.example.picket.common.exception.ErrorCode;
+import com.example.picket.domain.ticket.dto.response.GetTicketResponse;
+import com.example.picket.domain.ticket.entity.Ticket;
+import com.example.picket.domain.ticket.repository.TicketRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class TicketQueryService {
+
+    private final TicketRepository ticketRepository;
+
+    @Transactional(readOnly = true)
+    public Page<GetTicketResponse> getTickets(Long userId, int size, int page) {
+
+        Sort sortStandard = Sort.by("createdAt").descending();
+        Pageable pageable = PageRequest.of(page - 1, size, sortStandard);
+
+        return ticketRepository.findByUser(userId, pageable).map(GetTicketResponse::from);
+
+    }
+
+    @Transactional(readOnly = true)
+    public GetTicketResponse getTicket(Long ticketId) {
+
+        Ticket ticket = ticketRepository.findByTicketId(ticketId).orElseThrow(
+                () -> new CustomException(ErrorCode.TICKET_NOT_FOUND)
+        );
+        return GetTicketResponse.from(ticket);
+
+    }
+}
